\name{sampleNames}
\alias{sampleNames}
\docType{data}
\title{
An example of sample names for the Illumina 450k array
}
\description{
Sample names for a dataset containing 36 samples from the Illumina Infinium HumanMethylation 450k array.

}
\usage{data(sampleNames)}
\format{
A character vector 
}
\details{
Names are of the form 5621146023_R01C01. These sample names correspond to the names of the samples in \code{\link{betaMatrix}}
}
\seealso{
  \code{\link[ARRmNormalization]{getDesignInfo}} to obtain design information,
}
\examples{
data(sampleNames)
}
\keyword{datasets}
